// src/types/medusa.ts
export interface Product {
	title: string
	subtitle: string | null
	description: string | null
	handle: string | null
	is_giftcard: boolean
	status: ProductStatus
	images: Image[]
	thumbnail: string | null
	options: ProductOption[]
	variants: ProductVariant[]
	profile_id: string
	profile: ShippingProfile
	weight: number | null
	length: number | null
	height: number | null
	width: number | null
	hs_code: string | null
	origin_country: string | null
	mid_code: string | null
	material: string | null
	collection_id: string | null
	collection: ProductCollection
	type_id: string | null
	type: ProductType
	tags: ProductTag[]
	discountable: boolean
	external_id: string | null
	metadata: Record<string, unknown> | null
	sales_channels: SalesChannel[]
	beforeInsert
}